requirements.txt
openenv[core]==0.1.13
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2
pyyaml>=6.0.2
gradio>=4.44.0
openai>=1.40.0
